refactor: separate register to a component
