+++ src/components/Profile/Register/constants.ts
@@ -1,0 +1,2 @@
+export const registerFormId = 'register-form';
+export const profilePicPlaceholderURL = '/images/pfp-placeholder.svg';
